# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: techy Minecraft Server API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/profile:
    get:
      operationId: getMyProfile
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    put:
      operationId: updateMyProfile
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
  /users/{userId}:
    get:
      operationId: getUserById
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/{userId}/promote:
    post:
      operationId: promoteUser
      tags: [admin]
      summary: Promote user to staff rank
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PromoteUserBody"
      responses:
        "200":
          description: Promoted user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
  /announcements:
    get:
      operationId: listAnnouncements
      tags: [announcements]
      summary: List announcements
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Announcements
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnnouncementListResponse"
    post:
      operationId: createAnnouncement
      tags: [announcements]
      summary: Create announcement (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAnnouncementBody"
      responses:
        "201":
          description: Created announcement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
  /announcements/{announcementId}:
    get:
      operationId: getAnnouncement
      tags: [announcements]
      summary: Get announcement by ID
      parameters:
        - name: announcementId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Announcement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
    put:
      operationId: updateAnnouncement
      tags: [announcements]
      summary: Update announcement (admin)
      parameters:
        - name: announcementId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAnnouncementBody"
      responses:
        "200":
          description: Updated announcement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
    delete:
      operationId: deleteAnnouncement
      tags: [announcements]
      summary: Delete announcement (admin)
      parameters:
        - name: announcementId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /gallery:
    get:
      operationId: listGalleryImages
      tags: [gallery]
      summary: List gallery images
      responses:
        "200":
          description: Gallery images
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GalleryImage"
    post:
      operationId: createGalleryImage
      tags: [gallery]
      summary: Add gallery image (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGalleryImageBody"
      responses:
        "201":
          description: Created gallery image
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryImage"
  /gallery/{imageId}:
    delete:
      operationId: deleteGalleryImage
      tags: [gallery]
      summary: Delete gallery image (admin)
      parameters:
        - name: imageId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        clerkUserId:
          type: string
        username:
          type: string
        displayName:
          type: string
        minecraftUsername:
          type: string
        bio:
          type: string
        rank:
          type: string
          enum: [player, vip, mvp, moderator, admin, owner]
        isAdmin:
          type: boolean
        isBanned:
          type: boolean
        postCount:
          type: integer
        joinedAt:
          type: string
        avatarUrl:
          type: string
      required:
        - id
        - clerkUserId
        - username
        - rank
        - isAdmin
        - isBanned
        - postCount
        - joinedAt
    UpdateProfileBody:
      type: object
      properties:
        displayName:
          type: string
        minecraftUsername:
          type: string
        bio:
          type: string
    PromoteUserBody:
      type: object
      properties:
        rank:
          type: string
          enum: [player, vip, mvp, moderator, admin]
      required:
        - rank
    UserListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/UserProfile"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - users
        - total
        - page
        - totalPages
...
    Announcement:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        type:
          type: string
          enum: [general, important, update, event]
        authorName:
          type: string
        pinned:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - title
        - content
        - type
        - authorName
        - pinned
        - createdAt
    CreateAnnouncementBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        type:
          type: string
          enum: [general, important, update, event]
        pinned:
          type: boolean
      required:
        - title
        - content
        - type
    AnnouncementListResponse:
      type: object
      properties:
        announcements:
          type: array
          items:
            $ref: "#/components/schemas/Announcement"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - announcements
        - total
        - page
        - totalPages
...
    GalleryImage:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        uploadedBy:
          type: string
        createdAt:
          type: string
      required:
        - id
        - title
        - imageUrl
        - uploadedBy
        - createdAt
    CreateGalleryImageBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        imageUrl:
          type: string
      required:
        - title
        - imageUrl
...