# lib/api-zod/src/generated/types/createAnnouncementBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * techy Minecraft Server API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateAnnouncementBodyType } from "./createAnnouncementBodyType";

export interface CreateAnnouncementBody {
  title: string;
  content: string;
  type: CreateAnnouncementBodyType;
  pinned?: boolean;
}