# lib/api-zod/src/generated/types/userListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * techy Minecraft Server API
 * OpenAPI spec version: 0.1.0
 */
import type { UserProfile } from "./userProfile";

export interface UserListResponse {
  users: UserProfile[];
  total: number;
  page: number;
  totalPages: number;
}